KravRisikoscenarioAccordionContent har nå blitt splittet i LimitedReadonly og EditMode
