+++ apps/frontend/src/components/risikoscenario/kravRisikoscenario/KravRisikoscenarioAccordionContent/KravRisikoscenarioAccordionContentHeader.tsx
@@ -1,0 +1,9 @@
+import { Heading } from '@navikt/ds-react'
+
+export const KravRisikoscenarioAccordionContentHeader = () => (
+  <Heading level="3" size="small">
+    Følgende tiltak gjelder for dette risikoscenarioet
+  </Heading>
+)
+
+export default KravRisikoscenarioAccordionContentHeader
